props reorganized
